back-end/requirements.txt
annotated-types==0.7.0
blinker==1.8.2
Bottleneck==1.3.7
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
colorama==0.4.6
cryptography==42.0.8
deprecation==2.1.0
firebase-admin==6.5.0
Flask==3.0.3
functions-framework==3.7.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
numexpr==2.8.7
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pip==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
setuptools==69.5.1
six==1.16.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.1
websockets==10.4
Werkzeug==3.0.3
wheel==0.43.0
